got tesserae working again in python3

+++ tesserae/ui2/__init__.py
@@ -7,5 +7,5 @@ from shiboken2 import wrapInstance
 
 def getMayaWindow():
 	ptr = omui.MQtUtil.mainWindow()
-	widget = wrapInstance( long( ptr ), QtWidgets.QWidget )
+	widget = wrapInstance( int( ptr ), QtWidgets.QWidget )
 	return widget
